Rename Memphis CDC service file, update Memphis CDC and user servicesCo-authored-by: Memphis Dev <memphis@dev.com>
