+++ src/users/common/interface/erpHrV1User.dto.ts
@@ -1,5 +1,6 @@
 export interface ERPHRV1User {
   employeeID: string;
+  sr: string;
   name: string;
   firstName: string;
   middleName: string;
@@ -14,20 +15,6 @@ export interface ERPHRV1User {
   client: string;
   clientId: string;
   hrbpEmail: string;
-  employeeCompensation: Compensation;
-}
-
-interface Compensation {
-  basicMonthlySalary: number;
-  deminimis: number;
-  reimbursable: number;
-  taxableAllowance: number;
-  transpoAllowance: number;
-  parkingAllowance: number;
-  other: number;
-  discretionaryFirmBonus: number;
-  communicationAllowance: number;
-  medicalAllowance: number;
-  nightWorkIncentive: number;
-  phoneAllowance: number;
+  payrollBankAccount?: string | null;
+  payrollAccountNumber?: string | null;
 }
