+++ src/users/users.module.ts
@@ -1,14 +1,13 @@
 import { Module } from '@nestjs/common';
 import { UsersApiService } from './services/users.api.service';
 import { UsersController } from './users.controller';
-import { NewUserMemphisService } from './services/new-user.memphis.service';
-import { UpdateUserMemphisService } from './services/update-user.memphis.service';
+import { NewUserMemphisService } from './services/memphis/new-user.memphis.service';
+import { UpdateUserMemphisService } from './services/memphis/update-user.memphis.service';
 import { MemphisDevModule } from 'src/memphis-dev/memphis-dev.module';
 
 @Module({
   imports: [MemphisDevModule],
   controllers: [UsersController],
   providers: [UsersApiService, NewUserMemphisService, UpdateUserMemphisService],
-  exports: [NewUserMemphisService, UpdateUserMemphisService],
 })
 export class UsersModule {}
